Data Science and Cognitive Computing Courses
Objects in programming are like objects in real life. Like life, there are different classes of objects. In this notebook, we will create two classes called Circle and Rectangle. By the end of this notebook, you will have a better idea about :

what a class is
what an attribute is
what a method is